category changes

+++ lib/ApiFunctions/Category.tsx
@@ -1,28 +1,42 @@
+
+
 export type categoryType = {
     _id: string;
     name: string;
     img: string;
 }
 
+type ApiResponse = {
+    success: boolean,
+    message: string,
+    data?: categoryType[]
+}
+
 // category data
-export async function GetCategoryData(): Promise<categoryType[]> {
+export async function getCategoryData(): Promise<ApiResponse> {
 
-    const categoryAPI = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/categories`,
-        {
-            headers: { Accept: "application/json", },
-            method: "GET",
-            cache: "no-store",
-            next: { tags: ['category'] },
+    try {
+        const categoryAPI = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/categories`,
+            {
+                headers: { Accept: "application/json", },
+                method: "GET",
+                cache: "no-store",
+                next: { tags: ['category'] },
+            }
+        )
+    
+        if (!categoryAPI.ok) {
+            return { success: false, message: `Category API call failed with status ${categoryAPI.status}` }
         }
-    )
-
-    if (!categoryAPI.ok) {
-        throw new Error(`Category API call failed with status ${categoryAPI.status}`)
+    
+        const categoryData = await categoryAPI.json()
+    
+        return { success: true, message: "Brand Data found" , data: categoryData.data }
+        
+    } catch (error) {
+        return { success: false, message: `Something went wrong! : ${error}` }
     }
 
-    const categoryData = await categoryAPI.json()
-
-    return categoryData.data
 }
 
 // category ID data
@@ -47,19 +61,29 @@ export async function GetCategoryIdData(id: string): Promise<categoryType> {
 
 
 // DETETE Category By ID
-export async function DeleteCategoryById(id: string) {
+export async function deleteCategoryById(id: string): Promise<ApiResponse> {
 
-    const categoryAPI = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/categories/${id}`,
-        {
-            headers: { Accept: "application/json", },
-            method: "DELETE",
-            next: { tags: ['category'] },
-        }
-    )
-
-    if (!categoryAPI.ok) {
-        throw new Error(`Category ID API call failed with status ${categoryAPI.status}`)
+    if(!id) {
+        return { success: false, message: "Category ID is required." , }
     }
 
-    return categoryAPI.json()
+    try {
+        const categoryAPI = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/categories/${id}`,
+            {
+                headers: { Accept: "application/json", },
+                method: "DELETE",
+                next: { tags: ['category'] },
+            }
+        )
+    
+        if (!categoryAPI.ok) {
+            return { success: false, message: "Category not found." }
+        }
+
+        return { success: true, message: "Category Deleted Successfully" }
+
+    } catch (error) {
+        return { success: false, message: 'Something went wrong !' + error }
+    }
+
 }
